% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCovariate.R
\name{addCovariate}
\alias{addCovariate}
\title{Function}
\usage{
addCovariate(data = data, covariate, method = "nearest")
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples
}
