% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_cov_for_after_surveys.R
\name{adjust_cov_for_after_surveys}
\alias{adjust_cov_for_after_surveys}
\title{adjust_cov_for_after_surveys}
\usage{
adjust_cov_for_after_surveys(dt, end_date, cov, tier5, .x)
}
\arguments{
\item{dt}{is the date of ecological survey}

\item{end_date}{is the date associated with max disturbance values}

\item{cov}{refers to the extracted disturbance layer corresponding to the monitored years}

\item{tier5}{is the spatial unit}

\item{.x}{is the col from the disturbance data}
}
\description{
Adjust disturbance values
}
\details{
Adjusting disturbance values depending on survey dates
}
\author{
Julie Vercelloni
}
