% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_stage.R
\name{analysis_stage}
\alias{analysis_stage}
\title{Function}
\usage{
analysis_stage()
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
