% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.contrasts.lags.R
\name{annual.contrasts.lags}
\alias{annual.contrasts.lags}
\title{Fit model at tier level}
\usage{
annual.contrasts.lags(x, lags = 1)
}
\arguments{
\item{x}{A vector of values (e.g., years or time points).}

\item{lags}{Integer or vector of integers indicating the lag steps between values for contrasts. Default is 1.}
}
\value{
A matrix with rows corresponding to input values and columns to contrasts between lagged pairs.
}
\description{
Constructs a matrix representing annual contrasts with specified lags for a vector of values.
Each column corresponds to a contrast between year i and year i + lag.
}
\examples{
years <- 2010:2015
annual.contrasts.lags(years, lags = c(1,2))

}
\author{
Murray Logan
}
