% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_spatial_data.R
\name{assign_spatial_data}
\alias{assign_spatial_data}
\title{Assign spatial data}
\usage{
assign_spatial_data(data)
}
\arguments{
\item{data}{- a data frame containing the benthic data that
includes latidude/longitude}
}
\value{
dataframe - containing the data with tier classifications
}
\description{
Assign spatial data
}
\details{
This function assigns spatial domains to the benthic data. It does
so by first summarising the data at the site level and then before
assigning spatial domains to the data (based on the tiers sf
objects). Finally, these spatial inforation are joined onto the
benthic data
}
\author{
Murray
}
