% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignSpatialDomain_tier.R
\name{assignSpatialDomain_tier}
\alias{assignSpatialDomain_tier}
\title{Function}
\usage{
assignSpatialDomain_tier(dat, tier, andNearest = TRUE)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
