% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignSpatialDomain_tier2_plw.R
\name{assignSpatialDomain_tier2_plw}
\alias{assignSpatialDomain_tier2_plw}
\title{Function}
\usage{
assignSpatialDomain_tier2_plw(dat)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
