% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellmeans.R
\name{cellmeans}
\alias{cellmeans}
\title{Function}
\usage{
cellmeans(data.sub, GROUP, TIER, FOCAL_TIER)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples
}
