% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellmeans_reconstruct.R
\name{cellmeans_reconstruct}
\alias{cellmeans_reconstruct}
\title{Function}
\usage{
cellmeans_reconstruct(data.sub, GROUP, TIERS)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples
}
