% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellmeans_spatiotemporal.R
\name{cellmeans_spatiotemporal}
\alias{cellmeans_spatiotemporal}
\title{cellmeans_spatiotemporal}
\usage{
cellmeans_spatiotemporal(data.sub, GROUP, TIER)
}
\arguments{
\item{data.sub}{data set}

\item{GROUP}{Benthic group}

\item{TIERS}{tiers considered}
}
\description{
RawMeansSpatioTemporal
}
