% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellmeans_spatiotemporal_tier4.R
\name{cellmeans_spatiotemporal_tier4}
\alias{cellmeans_spatiotemporal_tier4}
\title{cellmeans_spatiotemporal_tier4}
\usage{
cellmeans_spatiotemporal_tier4(data.sub, GROUP, TIER)
}
\arguments{
\item{data.sub}{data set}

\item{GROUP}{Benthic group}

\item{TIERS}{tiers considered}
}
\description{
RawMeansSpatioTemporal at Tier 4 level
}
