% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellmeans_temporal.R
\name{cellmeans_temporal}
\alias{cellmeans_temporal}
\title{cellmeans_temporal}
\usage{
cellmeans_temporal(data.sub, GROUP, TIER)
}
\arguments{
\item{data.sub}{data set}

\item{GROUP}{Benthic group}

\item{TIERS}{tiers considered}
}
\description{
RawMeansTemporal
}
