% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_status.R
\name{change_status}
\alias{change_status}
\title{Function}
\usage{
change_status(stage, item, status, update_display = TRUE)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
