% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPackages.R
\name{checkPackages}
\alias{checkPackages}
\title{Function}
\usage{
checkPackages(log = TRUE)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
