% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoserver_cache.R
\name{clear_geoserver_cache}
\alias{clear_geoserver_cache}
\title{Clear Geoserver Cache}
\usage{
clear_geoserver_cache(tier_level = NULL)
}
\arguments{
\item{tier_level}{Optional tier level; NULL clears all tiers}
}
\value{
Number of files removed
}
\description{
Clear Geoserver Cache
}
