% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinedBoxes.R
\name{combinedBoxes}
\alias{combinedBoxes}
\title{Combine boxes}
\usage{
combinedBoxes(
  top,
  settings.box.text,
  main.box.text,
  bottom,
  box.width,
  settings.box.width,
  box.margins
)
}
\arguments{
\item{top}{top text}

\item{settings.box.text}{settings of box text}

\item{main.box.text}{main text}

\item{bottom}{bottom text}

\item{box.width}{box width}

\item{settings.box.width}{settings box width}

\item{box.margins}{box margins}
}
\value{
a vector of combined texts
}
\description{
combines boxes
}
