% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_predicates.R
\name{compare_predicates}
\alias{compare_predicates}
\title{Function}
\usage{
compare_predicates(PREDICATE, VALUE)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
