% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUnixToWindowsPath.R
\name{convertUnixToWindowsPath}
\alias{convertUnixToWindowsPath}
\title{Convert unix path to windows path}
\usage{
convertUnixToWindowsPath(unix_path)
}
\arguments{
\item{unix_path}{string of unix path}
}
\description{
converts unix-syntax path in windows path if operating in a windows environment
}
\examples{
convertUnixToWindowsPath("C:/Users/Documents/ReefCloud")
}
