% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_data_cut.R
\name{dev_data_cut}
\alias{dev_data_cut}
\title{Function}
\usage{
dev_data_cut(data, tier4)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
