% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endMatter.R
\name{endMatter}
\alias{endMatter}
\title{Function}
\usage{
endMatter()
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
