% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakeArgs.R
\name{fakeArgs}
\alias{fakeArgs}
\title{Function}
\usage{
fakeArgs(type)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
