% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_gaps.R
\name{fill_gaps}
\alias{fill_gaps}
\title{Fill data gaps}
\usage{
fill_gaps(data)
}
\arguments{
\item{data}{A dataframe (or tibble) containing the benthic data}
}
\value{
dataframe
}
\description{
Fill in data gaps
}
\details{
The data are record positive - that is, they only indicate what is
there and what is not there. For the purposes of the models, if a
benthic group is not encountered in a particular report year
within a specific sampling unit, then it should be recorded as 0.
So the following is intended to generate a full set benthic groups
within each spatio-temporal sampling unit...
}
\author{
Murray
}
