% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_focaltier.R
\name{filter_focaltier}
\alias{filter_focaltier}
\title{Filter Focal Tiers}
\usage{
filter_focaltier(data.grp, FOCAL_TIER)
}
\arguments{
\item{data.grp}{A data frame containing survey data to be filtered.}

\item{FOCAL_TIER}{A character string indicating the column name representing the tier.}
}
\value{
A list with two elements:
  \describe{
    \item{filtered_data}{The cleaned data frame with sufficient spatial and temporal coverage.}
    \item{removed_tiers}{A data frame of observations removed due to low coverage.}
  }
}
\description{
Filter FOCAL_TIERs with Insufficient Spatio-Temporal Coverage
}
\details{
Filters data based on minimum spatial and temporal observations per FOCAL_TIER.
Tiers with too few observations across space or years are removed and listed separately.
}
\examples{
\dontrun{
result <- filter_focaltier(data.grp, "Tier5")
filtered <- result$filtered_data
removed <- result$removed_tiers
}
}
\author{
Julie Vercelloni
}
