% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_non_collinear.R
\name{filter_non_collinear}
\alias{filter_non_collinear}
\title{Filter non-collinear variables}
\usage{
filter_non_collinear(df, vars, threshold = 0.7)
}
\arguments{
\item{df}{Data frame containing the variables to evaluate.}

\item{vars}{Character vector of variable names to test for collinearity.}

\item{threshold}{Numeric correlation threshold above which variables are considered collinear. Default is 0.7.}
}
\value{
Character vector of selected variables after filtering collinear predictors.
}
\description{
Identifies and removes highly correlated variables based on a correlation threshold, while retaining variables with the lowest temporal lag within predefined groups (e.g., cyc and dhw).
}
\examples{
selected_vars <- filter_non_collinear(
  df = my_data,
  vars = c("max_cyc_lag1", "max_cyc_lag3", "max_dhw_lag1"),
  threshold = 0.7
)
}
\author{
Julie Vercelloni
}
