% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraction_of_bbox.R
\name{fraction_of_bbox}
\alias{fraction_of_bbox}
\title{Function}
\usage{
fraction_of_bbox(bbox, frac = 0.25)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
