% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frk_prep_v2.R
\name{frk_prep_v2}
\alias{frk_prep_v2}
\title{frk_prep}
\usage{
frk_prep_v2(data.grp.tier, HexPred_reefid2, i, N, basis_resolution = 3L)
}
\arguments{
\item{data.grp.tier}{Data frame with observations for model fitting, including columns: LONGITUDE, LATITUDE, fYEAR, TOTAL, etc.}

\item{HexPred_reefid2}{Spatial polygons with covariates and reef IDs}
}
\value{
A list with prepared objects for FRK: ST_BAUs, STObj, basis functions
}
\description{
Prepare objects for the FRK model
}
\examples{
\dontrun{
# Example usage:
obj <- frk_prep(data.grp.tier = my_data, HexPred_reefid2 = my_spatial_covariates)
plot(obj$basis)
}
}
\author{
Julie Vercelloni
}
