% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSettings.R
\name{generateSettings}
\alias{generateSettings}
\title{Generate settings}
\usage{
generateSettings()
}
\value{
Creates the following global variables\cr
DATA_PATH\cr
INPUT_DATA\cr
RDATA_FILE \cr
FILENAME \cr
INPUT_FORMAT \cr
CSV_FILE \cr
TIER_DATA
}
\description{
Generates path and naming settings
}
\examples{
generateSettings()
}
