% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_tier_data.R
\name{get_cache_status}
\alias{get_cache_status}
\title{Get Cache Status}
\usage{
get_cache_status()
}
\value{
A character vector of cached object names
}
\description{
Returns information about what's currently cached
}
