% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_and_legacy_for_processing.R
\name{get_data_and_legacy_for_processing}
\alias{get_data_and_legacy_for_processing}
\title{Get Data and Legacy Data for Processing}
\usage{
get_data_and_legacy_for_processing()
}
\value{
A dataframe containing combined data and legacy data.
}
\description{
Loads the data and legacy data for processing and combines them into a single data frame.
}
\examples{
data <- get_data_and_legacy_for_processing()
head(data)
}
\author{
Murray Logan
}
