% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geoserver_info.R
\name{get_geoserver_info}
\alias{get_geoserver_info}
\title{Get Geoserver Info}
\usage{
get_geoserver_info()
}
\description{
Retrieves metadata from the geoserver.
This function loads available reefCloud layers and geoserver URL into the global environment.
}
\examples{
get_geoserver_info()
# This will load `rc_client` in the global environment
}
\author{
Murray Logan
}
