% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lag_cov.R
\name{get_lag_cov}
\alias{get_lag_cov}
\title{get_lag_cov}
\usage{
get_lag_cov(yr, cov, tier5, .x)
}
\arguments{
\item{yr}{is the adjusted year of the ecological survey}

\item{cov}{refers to the extracted disturbance layer corresponding to the monitored years}

\item{tier5}{is the spatial unit}

\item{.x}{is the col from the disturbance data}

\item{cov_name}{is the disturbance layer name}
}
\description{
Filter disturbance values to extract lags
}
\details{
Extract values of disturbance at previous time
}
\author{
Julie Vercelloni
}
