% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legacy_data.R
\name{get_legacy_data}
\alias{get_legacy_data}
\title{Get legacy data}
\usage{
get_legacy_data()
}
\value{
NULL (used for side effects: file copy, unzip, read, and save)
}
\description{
Unzip, retrieve, read, and save legacy benthic data for analysis and reporting
}
\examples{
get_legacy_data()
}
\author{
Murray
}
