% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stages.R
\name{get_stages}
\alias{get_stages}
\title{get_stages}
\usage{
get_stages()
}
\description{
The following function selects specific key files that mark the start of major analysis stages and thus determines the totalnumber of analysis stages.  The function returns a vector from 1 to the maximum number of stages.
}
