% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_status_name.R
\name{get_status_name}
\alias{get_status_name}
\title{Get Status Name from Status Tracking}
\usage{
get_status_name(stage, item)
}
\arguments{
\item{stage}{Integer. The stage number within the status tracking object.}

\item{item}{Character. The name of the item within the specified stage.}
}
\value{
A character string representing the status name associated with the given item and stage.
}
\description{
Retrieves the status name associated with a given stage and item from the status tracking object.
}
\examples{
\dontrun{
get_status_name(stage = 3, item = "load_benthic_data")
}
}
\author{
Murray Logan
}
