% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tiers.R
\name{get_tiers}
\alias{get_tiers}
\title{Get tiers}
\usage{
get_tiers()
}
\value{
NULL (saves tier spatial data for tiers 2 to 5)
}
\description{
Retrieve tier spatial layers from the geoserver or AWS, save locally as `.RData`, and register for reporting
}
\examples{
get_tiers()
}
\author{
Murray Logan
}
