% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_tier_data.R
\name{get_tiers_lookup}
\alias{get_tiers_lookup}
\title{Get Tiers Lookup (Cached)}
\usage{
get_tiers_lookup(force_reload = FALSE)
}
\arguments{
\item{force_reload}{If TRUE, forces reload from disk even if cached}
}
\value{
The tiers.lookup object
}
\description{
Loads tiers.lookup data from file or returns cached version if already loaded
}
