% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_benthic_data.R
\name{import_benthic_data}
\alias{import_benthic_data}
\title{Import benthic data}
\usage{
import_benthic_data()
}
\value{
A data frame with the benthic data
}
\description{
Import benthic data
}
\details{
Import benthic data from a CSV file
}
\author{
Murray
}
