% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_log.R
\name{initialise_log}
\alias{initialise_log}
\title{Function}
\usage{
initialise_log()
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
