% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isParent.R
\name{isParent}
\alias{isParent}
\title{Is Parent?}
\usage{
isParent()
}
\value{
TRUE if is parent, FALSE if not
}
\description{
The following function determines whether the current script is the ##
parent (directly called from command line etc) or child (sourced from another R script.
}
\examples{
isParent()
}
