% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_covariates.R
\name{lag_covariates}
\alias{lag_covariates}
\title{lag covariates}
\usage{
lag_covariates(cov, year_range, full_cov_lookup, cov_name)
}
\arguments{
\item{cov}{refers to the extracted disturbance layer corresponding to the monitored years}

\item{year_range}{corresponds to the range of monitoring years}

\item{full_cov_lookup}{is the full time series of the disturbance}

\item{cov_name}{is the disturbance layer name}
}
\description{
Compute lag values
}
\author{
Julie Vercelloni
}
