% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoserver_cache.R
\name{load_geoserver_cache}
\alias{load_geoserver_cache}
\title{Load Cached Geoserver Data}
\usage{
load_geoserver_cache(cov_name, tier_level)
}
\arguments{
\item{cov_name}{Coverage name}

\item{tier_level}{Tier level}
}
\value{
Cached object, or NULL on miss/error
}
\description{
Load Cached Geoserver Data
}
