% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_predictive_layers.R
\name{load_predictive_layers}
\alias{load_predictive_layers}
\title{Load Predictive Layers}
\usage{
load_predictive_layers(i, N)
}
\arguments{
\item{NULL}{No arguments required.}
}
\value{
A data object `full_cov_raw`, typically a `sf` or data frame with covariates.
}
\description{
Load Predictive Layers with Covariates
}
\details{
Loads spatial prediction layers containing environmental covariates for Tier 5.
}
\examples{
\dontrun{
full_cov <- load_predictive_layers()
}
}
\author{
Julie Vercelloni
}
