% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_utils.R
\name{log_pipeline_step}
\alias{log_pipeline_step}
\title{Log pipeline step with details}
\usage{
log_pipeline_step(step_num, step_name, details = list())
}
\arguments{
\item{step_num}{Step number}

\item{step_name}{Step description}

\item{details}{Named list of details to log}
}
\description{
Log pipeline step with details
}
