% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logBox.R
\name{logBox}
\alias{logBox}
\title{Function}
\usage{
logBox(box.width, box.margins)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
