% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainBox.R
\name{mainBox}
\alias{mainBox}
\title{Function}
\usage{
mainBox(settings.box.text, box.width, settings.box.width, box.margins)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples
}
