% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_reefid.R
\name{make_reefid}
\alias{make_reefid}
\title{Make Reef ID}
\usage{
make_reefid(tier.sf.joined, HexPred_sf, reef_layer.sf, i, N)
}
\arguments{
\item{tier.sf.joined}{A spatial layer with tier information (e.g., from `join_covariates_to_tier_lookup()`).}

\item{HexPred_sf}{Covariate shapefile for prediction.}

\item{reef_layer.sf}{A reef polygon spatial layer.}
}
\value{
A spatial `sf` object with Tier5 and reefid columns.
}
\description{
Make Reef ID
}
\examples{
\dontrun{
reefid_layer <- make_reefid(tier.sf.joined, HexPred_sf, reef_layer.sf)
}
}
\author{
Julie Vercelloni
}
