% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshINLA.R
\name{meshINLA}
\alias{meshINLA}
\title{Function}
\usage{
meshINLA(data.sub, tier5.sf, tier5_buffer.list, FOCAL_TIER)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples
}
