% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshINLA_cellmeans.R
\name{meshINLA_cellmeans}
\alias{meshINLA_cellmeans}
\title{Function}
\usage{
meshINLA_cellmeans(
  mesh,
  newdata.grid.list,
  spatial_and_stack.list,
  year = "fixed"
)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
