% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshINLA_data_grid.R
\name{meshINLA_data_grid}
\alias{meshINLA_data_grid}
\title{Function}
\usage{
meshINLA_data_grid(data.grp, tier5.sf, tier5_buffer.list)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
