% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshINLA_formula.R
\name{meshINLA_formula}
\alias{meshINLA_formula}
\title{Function}
\usage{
meshINLA_formula(obs.grid.list, spatial_and_stack.list, year = "fixed")
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
