% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshINLA_mesh.R
\name{meshINLA_mesh}
\alias{meshINLA_mesh}
\title{Function}
\usage{
meshINLA_mesh(full.grid.list, tier5_buffer.sf)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
