% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshINLA_newdata_grid.R
\name{meshINLA_newdata_grid}
\alias{meshINLA_newdata_grid}
\title{Function}
\usage{
meshINLA_newdata_grid(data.sub, FOCAL_TIER)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
