% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshINLA_obs_grid.R
\name{meshINLA_obs_grid}
\alias{meshINLA_obs_grid}
\title{Function}
\usage{
meshINLA_obs_grid(data.sub, tier5.sf, tier5_buffer.sf)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
