% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshINLA_reconstruct.R
\name{meshINLA_reconstruct}
\alias{meshINLA_reconstruct}
\title{Function}
\usage{
meshINLA_reconstruct(GROUP, TIERS)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
