% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshINLA_SPDEandStacks.R
\name{meshINLA_SPDEandStacks}
\alias{meshINLA_SPDEandStacks}
\title{Function}
\usage{
meshINLA_SPDEandStacks(mesh, full.grid.list, obs.grid.list, year = "fixed")
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
