% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshINLA_tier3.R
\name{meshINLA_tier3}
\alias{meshINLA_tier3}
\title{Function}
\usage{
meshINLA_tier3(cellmeans.full)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
