% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshINLA_tier3TemporalFocal.R
\name{meshINLA_tier3TemporalFocal}
\alias{meshINLA_tier3TemporalFocal}
\title{Function}
\usage{
meshINLA_tier3TemporalFocal(cellmeans.full.tier3)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
