% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshINLA_tier4.R
\name{meshINLA_tier4}
\alias{meshINLA_tier4}
\title{Function}
\usage{
meshINLA_tier4(cellmeans.full)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
