% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshINLA_tier5TemporalFocal.R
\name{meshINLA_tier5TemporalFocal}
\alias{meshINLA_tier5TemporalFocal}
\title{Function}
\usage{
meshINLA_tier5TemporalFocal(data.sub, cellmeans.full, FOCAL_TIER)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
