% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshINLA_tier_sub.R
\name{meshINLA_tier_sub}
\alias{meshINLA_tier_sub}
\title{Function}
\usage{
meshINLA_tier_sub(FOCAL_TIER, TIER, tier5.sf)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
