% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_quant_75.R
\name{mid_quant_75}
\alias{mid_quant_75}
\title{Compute the 75th Mid-Quantile of a Numeric Vector}
\usage{
mid_quant_75(y)
}
\arguments{
\item{y}{A numeric vector.}
}
\value{
A single numeric value representing the 75th mid-quantile of `y`.
}
\description{
Computes the 75th percentile (mid-quantile) of a numeric vector using the 
`midquantile()` function from the `Qtools` package.
The mid-quantile is useful for skewed data because it smooths
the empirical cumulative distribution function.
}
\examples{
\dontrun{
library(Qtools)
x <- c(0, 0, 1, 3, 5)
mid_quant_75(x)
}
}
\author{
Julie Vercelloni
}
