% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitmodel.R
\name{model_fitModel}
\alias{model_fitModel}
\title{Fit model}
\usage{
model_fitModel(raw_cell_means = TRUE)
}
\arguments{
\item{raw_cell_means}{Logical. If TRUE, fits raw cell means model. If FALSE, fits tier-level model. Default is TRUE.}
}
\description{
Fits the appropriate statistical model to benthic data, depending on the domain category.
If `DOMAIN_CATEGORY` is "site", it runs the site-level model; otherwise, it fits the tier-level model.
Optionally runs raw or smoothed cell means based on the `raw_cell_means` flag.
}
\examples{
# Fit the model using raw cell means
model_fitModel(raw_cell_means = TRUE)
}
\author{
Murray Logan
}
