% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitModelSite.R
\name{model_fitModelSite}
\alias{model_fitModelSite}
\title{Fit model at site level}
\usage{
model_fitModelSite(raw_cell_means = TRUE)
}
\arguments{
\item{raw_cell_means}{boolean. If TRUE, runs raw cell means. If FALSE, runs simple INLA cell means}
}
\description{
Fits model to data at site level, temporal model
}
\examples{
model_fitModelSite()
}
