% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitModelTier_type3.R
\name{model_fitModelTier_type3}
\alias{model_fitModelTier_type3}
\title{Fit model at tier level}
\usage{
model_fitModelTier_type3(data.grp)
}
\arguments{
\item{data.grp}{data on which model is fitted}
}
\description{
Fits model to data at tier level
}
\examples{
model_fitModelTier()
}
