% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitModelTier_type5.R
\name{model_fitModelTier_type5}
\alias{model_fitModelTier_type5}
\title{Fit model at tier level}
\usage{
model_fitModelTier_type5(data.grp, tier.sf)
}
\arguments{
\item{data.grp}{data on which model is fitted}

\item{covs.hexpred}{covariates shapefile}
}
\description{
Fits model to data at tier level
}
\examples{
model_fitModelTier()
}
