% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitModelTier_type5_v3.R
\name{model_fitModelTier_type5_v3}
\alias{model_fitModelTier_type5_v3}
\title{Fit model at tier level}
\usage{
model_fitModelTier_type5_v3(data.grp.enough, tier.sf)
}
\arguments{
\item{data.grp.enough}{Data frame with sufficient data for modelling (filtered upstream)}

\item{tier.sf}{Tier-level shapefile (e.g., Tier3 or Tier4 spatial scale)}
}
\description{
Fits FRK model to monitoring data at a given tier level, including covariate processing, quality control, model fitting and predictions.
}
\examples{
model_fitModelTier_type5_v3(my_data_filtered, my_tier_sf)
}
\author{
Julie Vercelloni
}
