% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitModelTier_v2.R
\name{model_fitModelTier_v2}
\alias{model_fitModelTier_v2}
\title{Fit model at tier level}
\usage{
model_fitModelTier_v2()
}
\description{
Fits spatio-temporal models to data at a tier level for benthic groups.
Selects and fits model types 1 to 6 based on `MODEL_TYPE` and scales up the predictions.
}
\examples{
model_fitModelTier_v2()

}
\author{
Julie Vercelloni
}
