% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_loadData_unix.R
\name{model_loadData_unix}
\alias{model_loadData_unix}
\title{Load Data from S3 or Local Sources}
\usage{
model_loadData_unix()
}
\arguments{
\item{None}{This function uses global settings and paths.}
}
\value{
NULL All datasets are saved to disk; nothing is returned.
}
\description{
Copies all required data sources from the bucket or local path into `<DATA_PATH>/primary`.\cr
- On the dev server (HPC), `DATA_FROM` must be `"LOCAL"`.\cr
- On the prod server (AWS), `DATA_FROM` must be `"S3"`.\cr
Required data collections under `<AWS_PATH>/raw` include:\cr
• Benthic data (`reefs_data.csv` or `.zip`)\cr
• Spatial tiers data (`tiers[2-5].shp`)\cr
• Covariates data (format TBD)
}
\examples{
model_loadData_unix()
}
\author{
Murray Logan
}
