% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_processDataSite.R
\name{model_processDataSite}
\alias{model_processDataSite}
\title{Process data at site scale}
\usage{
model_processDataSite()
}
\description{
## This script predominantly loads the benthic data (including legacy
}
\examples{
model_processData()
}
