% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_processDataTier.R
\name{model_processDataTier}
\alias{model_processDataTier}
\title{Process data at Tier scale}
\usage{
model_processDataTier()
}
\description{
## This script predominantly loads the benthic data (including legacy
}
\examples{
model_processData()
}
