% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summariseModelTier.R
\name{model_summariseModelTier}
\alias{model_summariseModelTier}
\title{Summarise model at tier level}
\usage{
model_summariseModelTier()
}
\description{
Summarise model at tier level
}
\examples{
model_summariseModelTier()
}
