% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openingBanner.R
\name{openingBanner}
\alias{openingBanner}
\title{Function}
\usage{
openingBanner()
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
