% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openingBanner1.R
\name{openingBanner1}
\alias{openingBanner1}
\title{Function}
\usage{
openingBanner1()
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
