% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outerBox.bottom.R
\name{outerBox.bottom}
\alias{outerBox.bottom}
\title{Function}
\usage{
outerBox.bottom(outer.box.width, this.box.width)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples
}
