% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{prepare_covariates}
\alias{prepare_covariates}
\title{Prepare covariates}
\usage{
prepare_covariates(data)
}
\arguments{
\item{data}{- the benthic data}
}
\description{
Prepare covariates
}
\details{
Prepare covariates and join to the benthic data
}
\author{
Murray
}
