% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for analyses}
\usage{
prepare_data(data)
}
\arguments{
\item{data}{- a data frame containing the benthic data}
}
\value{
dataframe
}
\description{
Prepare data for analyses
}
\details{
Prepare data for analyses
}
\author{
Murray
}
