% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_status.R
\name{read_status}
\alias{read_status}
\title{Function}
\usage{
read_status()
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
