% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud__add_stage.R
\name{ReefCloud__add_stage}
\alias{ReefCloud__add_stage}
\title{Function}
\usage{
ReefCloud__add_stage(stage, title)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
