% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud__add_status.R
\name{ReefCloud__add_status}
\alias{ReefCloud__add_status}
\title{Function}
\usage{
ReefCloud__add_status(stage, item, label, status, update_display = TRUE)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
