% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud__logBox.R
\name{ReefCloud__logBox}
\alias{ReefCloud__logBox}
\title{Function}
\usage{
ReefCloud__logBox(box.width, box.margins)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
