% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud__mainBox.R
\name{ReefCloud__mainBox}
\alias{ReefCloud__mainBox}
\title{Function}
\usage{
ReefCloud__mainBox(
  settings.box.text,
  box.width,
  settings.box.width,
  box.margins
)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
