% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud__read_status.R
\name{ReefCloud__read_status}
\alias{ReefCloud__read_status}
\title{Function}
\usage{
ReefCloud__read_status()
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
