% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud__remove_predicates.R
\name{ReefCloud__remove_predicates}
\alias{ReefCloud__remove_predicates}
\title{Function}
\usage{
ReefCloud__remove_predicates(update_display = TRUE)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
