% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud__remove_status.R
\name{ReefCloud__remove_status}
\alias{ReefCloud__remove_status}
\title{Function}
\usage{
ReefCloud__remove_status(stage, item)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
