% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_addCovariate.R
\name{ReefCloud_addCovariate}
\alias{ReefCloud_addCovariate}
\title{Function}
\usage{
ReefCloud_addCovariate(data = data, covariate, method = "nearest")
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
