% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_assignSpatialDomain_tier.R
\name{ReefCloud_assignSpatialDomain_tier}
\alias{ReefCloud_assignSpatialDomain_tier}
\title{Function}
\usage{
ReefCloud_assignSpatialDomain_tier(dat, tier, andNearest = TRUE)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
