% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_assignSpatialDomain_tier5_plw.R
\name{ReefCloud_assignSpatialDomain_tier5_plw}
\alias{ReefCloud_assignSpatialDomain_tier5_plw}
\title{Function}
\usage{
ReefCloud_assignSpatialDomain_tier5_plw(dat)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
