% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_build_report.R
\name{ReefCloud_build_report}
\alias{ReefCloud_build_report}
\title{Function}
\usage{
ReefCloud_build_report(component = NULL, string = NULL)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
