% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_cellmeans.R
\name{ReefCloud_cellmeans}
\alias{ReefCloud_cellmeans}
\title{Function}
\usage{
ReefCloud_cellmeans(data.sub, GROUP, TIER, FOCAL_TIER)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
