% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_cellmeans_reconstruct.R
\name{ReefCloud_cellmeans_reconstruct}
\alias{ReefCloud_cellmeans_reconstruct}
\title{Function}
\usage{
ReefCloud_cellmeans_reconstruct(data.sub, GROUP, TIERS)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
