% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_cellmeans_spatiotemporal_tier3.R
\name{ReefCloud_cellmeans_spatiotemporal_tier3}
\alias{ReefCloud_cellmeans_spatiotemporal_tier3}
\title{Function}
\usage{
ReefCloud_cellmeans_spatiotemporal_tier3(data.sub, GROUP, TIER)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
