% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_checkPackages.R
\name{ReefCloud_checkPackages}
\alias{ReefCloud_checkPackages}
\title{Function}
\usage{
ReefCloud_checkPackages(log = TRUE)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
