% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_endMatter.R
\name{ReefCloud_endMatter}
\alias{ReefCloud_endMatter}
\title{Function}
\usage{
ReefCloud_endMatter()
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
