% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_fake.R
\name{ReefCloud_fake}
\alias{ReefCloud_fake}
\title{Function}
\usage{
ReefCloud_fake()
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
