% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_fakeArgs.R
\name{ReefCloud_fakeArgs}
\alias{ReefCloud_fakeArgs}
\title{Function}
\usage{
ReefCloud_fakeArgs(type)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
