% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_generateSettings.R
\name{ReefCloud_generateSettings}
\alias{ReefCloud_generateSettings}
\title{Function}
\usage{
ReefCloud_generateSettings()
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
