% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_get_stages.R
\name{ReefCloud_get_stages}
\alias{ReefCloud_get_stages}
\title{Function}
\usage{
ReefCloud_get_stages()
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
