% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_initialise_status.R
\name{ReefCloud_initialise_status}
\alias{ReefCloud_initialise_status}
\title{Function}
\usage{
ReefCloud_initialise_status()
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
