% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_isParent.R
\name{ReefCloud_isParent}
\alias{ReefCloud_isParent}
\title{Is Parent?}
\usage{
ReefCloud_isParent()
}
\value{
TRUE if is parent, FALSE if not
}
\description{
The following function determines whether the current script is the ##
parent (directly called from command line etc) or child (sourced from another R script.
}
\examples{
ReefCloud_isParent()
}
