% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_log.R
\name{ReefCloud_log}
\alias{ReefCloud_log}
\title{Function}
\usage{
ReefCloud_log(status, logFile, Category, msg = NULL)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
