% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_make_tiers_figures.R
\name{ReefCloud_make_tiers_figures}
\alias{ReefCloud_make_tiers_figures}
\title{Function}
\usage{
ReefCloud_make_tiers_figures(tier.sf)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
