% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_mean_median_hdci.R
\name{ReefCloud_mean_median_hdci}
\alias{ReefCloud_mean_median_hdci}
\title{Function}
\usage{
ReefCloud_mean_median_hdci(.data, ...)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
