% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_meshINLA.R
\name{ReefCloud_meshINLA}
\alias{ReefCloud_meshINLA}
\title{Function}
\usage{
ReefCloud_meshINLA(data.sub, tier5.sf, tier5_buffer.list, FOCAL_TIER)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
