% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_meshINLA_data_grid.R
\name{ReefCloud_meshINLA_data_grid}
\alias{ReefCloud_meshINLA_data_grid}
\title{Function}
\usage{
ReefCloud_meshINLA_data_grid(data.grp, tier5.sf, tier5_buffer.list)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
