% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_meshINLA_fit.R
\name{ReefCloud_meshINLA_fit}
\alias{ReefCloud_meshINLA_fit}
\title{Function}
\usage{
ReefCloud_meshINLA_fit(form, spatial_and_stack.list)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
