% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_meshINLA_formula.R
\name{ReefCloud_meshINLA_formula}
\alias{ReefCloud_meshINLA_formula}
\title{Function}
\usage{
ReefCloud_meshINLA_formula(
  obs.grid.list,
  spatial_and_stack.list,
  year = "fixed"
)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
