% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_meshINLA_full_grid.R
\name{ReefCloud_meshINLA_full_grid}
\alias{ReefCloud_meshINLA_full_grid}
\title{Function}
\usage{
ReefCloud_meshINLA_full_grid(data.sub, tier5_buffer.sf)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
