% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_meshINLA_mesh.R
\name{ReefCloud_meshINLA_mesh}
\alias{ReefCloud_meshINLA_mesh}
\title{Function}
\usage{
ReefCloud_meshINLA_mesh(full.grid.list, tier5_buffer.sf)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
