% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_meshINLA_newdata_grid.R
\name{ReefCloud_meshINLA_newdata_grid}
\alias{ReefCloud_meshINLA_newdata_grid}
\title{Function}
\usage{
ReefCloud_meshINLA_newdata_grid(data.sub, FOCAL_TIER)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
