% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_meshINLA_SPDEandStacks.R
\name{ReefCloud_meshINLA_SPDEandStacks}
\alias{ReefCloud_meshINLA_SPDEandStacks}
\title{Function}
\usage{
ReefCloud_meshINLA_SPDEandStacks(
  mesh,
  full.grid.list,
  obs.grid.list,
  year = "fixed"
)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
