% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_meshINLA_tier2.R
\name{ReefCloud_meshINLA_tier2}
\alias{ReefCloud_meshINLA_tier2}
\title{Function}
\usage{
ReefCloud_meshINLA_tier2(cellmeans.full)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
