% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_meshINLA_tier2Spatiotemporal.R
\name{ReefCloud_meshINLA_tier2Spatiotemporal}
\alias{ReefCloud_meshINLA_tier2Spatiotemporal}
\title{Function}
\usage{
ReefCloud_meshINLA_tier2Spatiotemporal(cellmeans.full.tier2)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
