% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_meshINLA_tier3Spatiotemporal.R
\name{ReefCloud_meshINLA_tier3Spatiotemporal}
\alias{ReefCloud_meshINLA_tier3Spatiotemporal}
\title{Function}
\usage{
ReefCloud_meshINLA_tier3Spatiotemporal(cellmeans.full.tier3)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
