% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_meshINLA_tier4Spatiotemporal.R
\name{ReefCloud_meshINLA_tier4Spatiotemporal}
\alias{ReefCloud_meshINLA_tier4Spatiotemporal}
\title{Function}
\usage{
ReefCloud_meshINLA_tier4Spatiotemporal(cellmeans.full.tier4)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
