% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_meshINLA_tier_sub.R
\name{ReefCloud_meshINLA_tier_sub}
\alias{ReefCloud_meshINLA_tier_sub}
\title{Function}
\usage{
ReefCloud_meshINLA_tier_sub(FOCAL_TIER, TIER, tier5.sf)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
