% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_openingBanner.R
\name{ReefCloud_openingBanner}
\alias{ReefCloud_openingBanner}
\title{Function}
\usage{
ReefCloud_openingBanner()
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
