% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_openingBanner1.R
\name{ReefCloud_openingBanner1}
\alias{ReefCloud_openingBanner1}
\title{Function}
\usage{
ReefCloud_openingBanner1()
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
