% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_parseCLA.R
\name{ReefCloud_parseCLA}
\alias{ReefCloud_parseCLA}
\title{Function}
\usage{
ReefCloud_parseCLA(args)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
