% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_simpleINLA_data.R
\name{ReefCloud_simpleINLA_data}
\alias{ReefCloud_simpleINLA_data}
\title{Function}
\usage{
ReefCloud_simpleINLA_data(data.sub, GROUP, TIER)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
