% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_simpleINLA_tier3.R
\name{ReefCloud_simpleINLA_tier3}
\alias{ReefCloud_simpleINLA_tier3}
\title{Function}
\usage{
ReefCloud_simpleINLA_tier3(data.sub, GROUP, .data.tier5)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
