% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_simpleINLA_tier5a.R
\name{ReefCloud_simpleINLA_tier5a}
\alias{ReefCloud_simpleINLA_tier5a}
\title{Function}
\usage{
ReefCloud_simpleINLA_tier5a(data.sub, GROUP, TIER, .data, .data.fit)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
