% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_startMatter.R
\name{ReefCloud_startMatter}
\alias{ReefCloud_startMatter}
\title{Function}
\usage{
ReefCloud_startMatter(args = commandArgs())
}
\arguments{
\item{args}{command line arguments (defaults commandArgs())}
}
\value{
returned arguments description
}
\description{
The following function is a wrapper around a number of routines
that occur once at the start of an analysis run. In order, these routines are:
- initialise a status list.  This list is the basis of information presented in the command line interface.
- initialise the log file
}
\examples{
ReefCloud_startMatter(args = c("--bucket=/home/data/AUS", "--domain=tier", "--by_tier=4", "--debug=true", "--runStage=1", "--refresh_data=false"))
}
