% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_tryCatch.R
\name{ReefCloud_tryCatch}
\alias{ReefCloud_tryCatch}
\title{tryCatch function for the reefCloudPackage}
\usage{
ReefCloud_tryCatch(
  expr,
  logFile,
  Category,
  expectedClass = NULL,
  msg = NULL,
  return = NULL,
  showWarnings = FALSE,
  stage = NULL,
  item = NULL
)
}
\arguments{
\item{expr}{expression to be evaluated.}

\item{logFile}{path for log file}

\item{Category}{ask}

\item{expectedClass}{ask}

\item{msg}{message}

\item{return}{ask}

\item{showWarnings}{Bool. Show warnings (TRUE) or not (FALSE).}

\item{stage}{ask}

\item{item}{ask}
}
\value{
updates Log file according to
}
\description{
des
}
