% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_whratio.R
\name{ReefCloud_whratio}
\alias{ReefCloud_whratio}
\title{Function}
\usage{
ReefCloud_whratio(object.sf)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
