% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_predicates.R
\name{remove_predicates}
\alias{remove_predicates}
\title{Function}
\usage{
remove_predicates(update_display = TRUE)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
