% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_status.R
\name{remove_status}
\alias{remove_status}
\title{Function}
\usage{
remove_status(stage, item)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
