% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_factor.R
\name{rm_factor}
\alias{rm_factor}
\title{Remove Factor Variables with Fewer than Two Levels}
\usage{
rm_factor(model_formula, data.sub)
}
\arguments{
\item{model_formula}{A model formula containing fixed and random effects.}

\item{data.sub}{A data frame with the variables used in the model.}
}
\value{
A cleaned model formula excluding low-level factors.
}
\description{
Cleans a model formula by removing factor variables that have fewer than two levels.
Continuous (numeric) variables are retained. Also removes random effect terms of the form \code{f(var)}.
}
\examples{
\dontrun{
data.sub <- data.frame(
  y = rnorm(10),
  factor1 = factor(rep("A", 10)),
  factor2 = factor(rep(c("A", "B"), 5)),
  numeric1 = rnorm(10)
)
model_formula <- y ~ factor1 + factor2 + numeric1 + f(factor2, model = 'iid')
rm_factor(model_formula, data.sub)
}

}
