% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_system.R
\name{safe_copy}
\alias{safe_copy}
\title{Safely Copy File}
\usage{
safe_copy(from, to, recursive = FALSE)
}
\arguments{
\item{from}{Source file path}

\item{to}{Destination file path}

\item{recursive}{If TRUE, copy directories recursively}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Copies a file using processx instead of system("cp ...")
}
