% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_system.R
\name{safe_tail}
\alias{safe_tail}
\title{Safely Read Log Tail}
\usage{
safe_tail(file, n = 5)
}
\arguments{
\item{file}{Path to file}

\item{n}{Number of lines to read (default: 5)}
}
\value{
Character vector of lines
}
\description{
Reads last N lines of a file using processx instead of system("tail ...")
}
