% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_benthic_data.R
\name{save_benthic_data}
\alias{save_benthic_data}
\title{Save benthic data}
\usage{
save_benthic_data(data)
}
\description{
Save benthic data
}
\details{
Save benthic data to RData file
}
\author{
Murray
}
