% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoserver_cache.R
\name{save_geoserver_cache}
\alias{save_geoserver_cache}
\title{Save Geoserver Data to Cache}
\usage{
save_geoserver_cache(data, cov_name, tier_level)
}
\arguments{
\item{data}{Object to cache}

\item{cov_name}{Coverage name}

\item{tier_level}{Tier level}
}
\value{
Invisible NULL
}
\description{
Save Geoserver Data to Cache
}
