% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_status.R
\name{save_status}
\alias{save_status}
\title{Save status}
\usage{
save_status()
}
\value{
save STATUS IN in DATA_PATH/STATUS.RData file\cr
and current values of settings in DATA_PATH/global_settings.RData
}
\description{
save current status and current global variables
}
\examples{
save_status()
}
