% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_up_pred_all.R
\name{scale_up_pred_all}
\alias{scale_up_pred_all}
\title{Scale Up Predictions from Model Outputs}
\usage{
scale_up_pred_all(whichModel)
}
\arguments{
\item{whichModel}{Character string indicating the model type (e.g., "type6").}
}
\value{
This function has no return value. It writes summarised prediction CSV files to AWS path.
}
\description{
This function reads model outputs (either type6 or other specified models), processes them
to scale up predictions across different tiers, extract covariate effect sizes, compute annual contrasts and saves the summarised results at all tiers.
}
\examples{
scale_up_pred_all("type6")
}
\author{
Julie Vercelloni
}
