% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_covariates.R
\name{select_covariates}
\alias{select_covariates}
\title{Select Spatially Representative Covariates}
\usage{
select_covariates(x, i, N)
}
\arguments{
\item{x}{A `sf` dataframe (e.g., `HexPred_sf`) containing covariates.}
}
\value{
A character vector of selected covariate names.
}
\description{
Selects covariates from a spatial dataset if their 70th percentile value is greater than 0,
assuming such variables are spatially representative.
}
\examples{
\dontrun{
library(sf)
library(dplyr)

# Example mock data
hex_grid <- st_sf(
  geometry = st_sfc(st_point(c(1, 2)), st_point(c(3, 4))),
  max_temp = c(0.1, 0.5),
  max_salinity = c(0, 0),
  max_wave = c(1.2, 1.5)
)

selected_vars <- select_covariates(hex_grid)
print(selected_vars)
}
}
\author{
Julie Vercelloni
}
