% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_lowest_lag.R
\name{select_lowest_lag}
\alias{select_lowest_lag}
\title{Select variable with lowest lag}
\usage{
select_lowest_lag(var1, var2, group1, group2)
}
\arguments{
\item{var1}{Character string representing the first variable name.}

\item{var2}{Character string representing the second variable name.}

\item{group1}{Character string indicating the group of the first variable.}

\item{group2}{Character string indicating the group of the second variable.}
}
\value{
A character vector containing the selected variable(s).
}
\description{
Compares two correlated variables within the same predictor group and retains the variable with the lowest temporal lag. If variables belong to different groups, both are retained.
}
\examples{
select_lowest_lag(
  var1 = "max_cyc_lag1",
  var2 = "max_cyc_lag3",
  group1 = "max_cyc",
  group2 = "max_cyc"
)
}
\author{
Julie Vercelloni
}
