% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settingsBox.R
\name{settingsBox}
\alias{settingsBox}
\title{Function}
\usage{
settingsBox(
  settings,
  box.width,
  box.nchar,
  box.margins,
  currentTime,
  max.settings.width
)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
