% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleINLA.R
\name{simpleINLA}
\alias{simpleINLA}
\title{Function}
\usage{
simpleINLA(data.sub, GROUP, TIER)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples
}
