% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleINLA_fit.R
\name{simpleINLA_fit}
\alias{simpleINLA_fit}
\title{Function}
\usage{
simpleINLA_fit(data.sub, GROUP, TIER, .data)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples
}
