% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleINLA_reconstruct.R
\name{simpleINLA_reconstruct}
\alias{simpleINLA_reconstruct}
\title{Function}
\usage{
simpleINLA_reconstruct(data.sub, GROUP, TIERS)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
