% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleINLA_tier5.R
\name{simpleINLA_tier5}
\alias{simpleINLA_tier5}
\title{Function}
\usage{
simpleINLA_tier5(data.sub, GROUP, TIER, .data, .data.fit.cm = .data.fit.cm)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
