% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startMatter1.R
\name{startMatter1}
\alias{startMatter1}
\title{Function}
\usage{
startMatter1()
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
