% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_years_from_predictive_layers.R
\name{trim_years_from_predictive_layers}
\alias{trim_years_from_predictive_layers}
\title{Trim years of the predictive layers}
\usage{
trim_years_from_predictive_layers(full_cov_raw, data.grp.tier)
}
\arguments{
\item{full_cov_raw}{covariates}

\item{data.grp.tier}{data grouped}
}
\value{
full_cov
}
\description{
Cut years associated with ecological survey
}
