% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_benthic_data.R
\name{unzip_benthic_data}
\alias{unzip_benthic_data}
\title{Unzip benthic data}
\usage{
unzip_benthic_data()
}
\value{
NULL (used for side effects: unzipping files)
}
\description{
Unzips the benthic data file to the primary data directory using system unzip
}
\examples{
unzip_benthic_data()
}
\author{
Murray Logan
}
