% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_formula.R
\name{update_formula}
\alias{update_formula}
\title{Remove a variable from model formula}
\usage{
update_formula(remove_var, formula_no_random)
}
\arguments{
\item{remove_var}{Character name of the variable to remove from the formula.}
}
\value{
A cleaned model formula string with the specified variable removed and formatting adjusted.
}
\description{
Automatically removes a specified covariate from a model formula string by matching typical additive patterns.
}
\author{
Julie Vercelloni
}
