% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_benthic_data.R
\name{validate_benthic_data}
\alias{validate_benthic_data}
\title{Validate benthic data}
\usage{
validate_benthic_data(data, rules)
}
\arguments{
\item{data}{Benthic data}

\item{rules}{Rules to validate the data}
}
\value{
Validation result or the violating rows of the benthic data
}
\description{
Validate the input benthic data against a set of predefined rules
}
\examples{
rules <- validator(!is.na(SITE_ID), COVER >= 0)
validate_benthic_data(benthic_data, rules)
}
\author{
Murray Logan
}
