% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_data}
\alias{validate_data}
\title{Validate data against rules}
\usage{
validate_data(data, rules)
}
\arguments{
\item{data}{A data frame to be validated.}

\item{rules}{A set of validation rules created with `validator()`.}
}
\value{
A tidy data frame summarising rule descriptions and validation results. Stops with an error if validation fails.
}
\description{
Validate a data frame using predefined rules, save failing rows if any, and return a tidy summary.
}
\examples{
rules <- validator(!is.na(SITE_ID), COVER >= 0)
validate_data(benthic_data, rules)
}
\author{
Murray Logan
}
