% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violating_rows.R
\name{violating_rows}
\alias{violating_rows}
\title{Get violating rows from validation result}
\usage{
violating_rows(result)
}
\arguments{
\item{result}{The result object returned by `validate::confront()`.}
}
\value{
A data frame of rows violating rules, or NULL if no violations are found.
}
\description{
Identify and return rows in the data that violate specified validation rules.
}
\examples{
rules <- validator(!is.na(SITE_ID), COVER >= 0)
result <- confront(benthic_data, rules)
violating_rows(result)
}
\author{
Murray Logan
}
