% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whratio.R
\name{whratio}
\alias{whratio}
\title{Function}
\usage{
whratio(object.sf)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
