% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_aws.R
\name{write_aws}
\alias{write_aws}
\title{Function}
\usage{
write_aws(file, level)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
