% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cov_to_data.R
\name{add_cov_to_data}
\alias{add_cov_to_data}
\title{Add covariates to data}
\usage{
add_cov_to_data(data, cov, cov_name)
}
\arguments{
\item{data}{Data on which model is fitted}

\item{cov}{Extracted disturbance layer corresponding to the monitored years}

\item{cov_name}{Name of the disturbance layer}
}
\value{
Data with adjusted disturbance values
}
\description{
Process spatial layers including time lags and adjusting values depending on survey dates
}
\examples{
examples
}
\author{
Julie Vercelloni
}
