% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_status.R
\name{add_status}
\alias{add_status}
\title{Function}
\usage{
add_status(stage, item, label, status, update_display = TRUE)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples
}
