% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_stage.R
\name{analysis_stage}
\alias{analysis_stage}
\title{Analysis stage}
\usage{
analysis_stage()
}
\value{
returned arguments description
}
\description{
Read in the stage that the analysis is up to
}
\examples{
examples
}
