% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_spatial_data.R
\name{assign_spatial_data}
\alias{assign_spatial_data}
\title{Assign Spatial Data}
\usage{
assign_spatial_data(data)
}
\arguments{
\item{data}{A data frame containing benthic data including latitude and longitude.}
}
\value{
A data frame with spatial tier classifications assigned.
}
\description{
Assigns spatial domain (tiers) to the benthic data. First, it summarises data at the site level, then assigns spatial domains using tier shapefiles, and finally joins the spatial info back to the full dataset.
}
\examples{
data_with_spatial <- assign_spatial_data(benthic_data)
head(data_with_spatial)
}
\author{
Murray Logan
}
