% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellmeans.R
\name{cellmeans}
\alias{cellmeans}
\title{Cellmeans}
\usage{
cellmeans(data.sub, GROUP, TIER, FOCAL_TIER)
}
\arguments{
\item{data.sub}{data set}

\item{GROUP}{benthic group}

\item{TIER}{tier}

\item{FOCAL_TIER}{tier lever on which predictions are made}
}
\description{
Modelling functions (tiers) using cellmeans
}
