% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellmeans_reconstruct.R
\name{cellmeans_reconstruct}
\alias{cellmeans_reconstruct}
\title{cellmeans_reconstruct}
\usage{
cellmeans_reconstruct(data.sub, GROUP, TIERS)
}
\arguments{
\item{data.sub}{data set}

\item{GROUP}{Benthic group}

\item{TIERS}{tiers considered}
}
\description{
Description
}
