% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellmeans_spatiotemporal_tier3.R
\name{cellmeans_spatiotemporal_tier3}
\alias{cellmeans_spatiotemporal_tier3}
\title{cellmeans_spatiotemporal_tier3}
\usage{
cellmeans_spatiotemporal_tier3(data.sub, GROUP, TIER)
}
\arguments{
\item{data.sub}{data set}

\item{GROUP}{Benthic group}

\item{TIERS}{tiers considered}
}
\description{
RawMeansSpatioTemporal at Tier 3 level
}
