% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellmeans_spatiotemporal_tier4.R
\name{cellmeans_spatiotemporal_tier4}
\alias{cellmeans_spatiotemporal_tier4}
\title{Function}
\usage{
cellmeans_spatiotemporal_tier4(data.sub, GROUP, TIER)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
