% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_status.R
\name{change_status}
\alias{change_status}
\title{Modify the STATUS global variable}
\usage{
change_status(stage, item, status, update_display = TRUE)
}
\arguments{
\item{stage}{current running stage}

\item{item}{name of item to be changed}

\item{status}{new value assigned to STATUS[[stage]]$item}

\item{update_display}{Bool: if TRUE prompts update in console. Default TRUE}
}
\value{
void. Modifies STATUS only
}
\description{
This function modifies the STATUS global variable
}
\examples{
change_status(stage = "SETTINGS", item = "DEBUG_MODE",
status = "success", update_display = FALSE)
}
