% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPackages.R
\name{checkPackages}
\alias{checkPackages}
\title{Check packages}
\usage{
checkPackages(log = TRUE)
}
\arguments{
\item{log}{= TRUE description}
}
\value{
returned arguments description
}
\description{
The following function checks to ensure that all the required packages are available on the system.
}
\examples{
examples
}
