% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraction_of_bbox.R
\name{fraction_of_bbox}
\alias{fraction_of_bbox}
\title{fraction_of_bbox}
\usage{
fraction_of_bbox(bbox, frac = 0.25)
}
\description{
The following function is used to return a fraction of the longest of the eight and width of a bouding box. This is useful if for when we want to havea buffer around a tier of (for example) 10% (frac = 0.1)
}
