% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coral_reef_shapefiles.R
\name{get_coral_reef_shape_files}
\alias{get_coral_reef_shape_files}
\title{Get coral reef shapefiles}
\usage{
get_coral_reef_shape_files()
}
\value{
NULL (saves shapefile as reef_layer.sf.RData)
}
\description{
Unzips and loads the global coral reef shapefiles from the reefCloud package, saves them locally, and removes temporary files
}
\examples{
get_coral_reef_shape_files()
}
\author{
Murray Logan
}
