% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covariates.R
\name{get_covariates}
\alias{get_covariates}
\title{Get covariates}
\usage{
get_covariates()
}
\description{
Get covariates
}
\details{
This function gets the covariates from the geoserver
}
\author{
Murray
}
