% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_and_legacy_for_processing.R
\name{get_data_and_legacy_for_processing}
\alias{get_data_and_legacy_for_processing}
\title{Get data and legacy data for processing}
\usage{
get_data_and_legacy_for_processing()
}
\value{
dataframe with data and legacy data
}
\description{
Get data and legacy data for processing
}
\details{
This function loads the data and legacy data for processing
and combines them into a single data frame.
}
\author{
Murray
}
