% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geoserver_data.R
\name{get_geoserver_data}
\alias{get_geoserver_data}
\title{Get data from geoserver}
\usage{
get_geoserver_data(Tier = 4, cov_name = NULL)
}
\arguments{
\item{Tier}{- an integer indicating the tier of the data to be retrieved}

\item{cov_name}{- a string indicating the name of the coverage to be retrieved}
}
\value{
a spatial data frame containing the data from the geoserver
}
\description{
Get data from geoserver
}
\details{
This function gets data from the geoserver
}
\author{
Murray
}
