% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geoserver_data.R
\name{get_geoserver_data}
\alias{get_geoserver_data}
\title{Get Data from Geoserver}
\usage{
get_geoserver_data(Tier = as.numeric(BY_TIER) - 1, cov_name = NULL)
}
\arguments{
\item{Tier}{An integer indicating the tier of the data to be retrieved.}

\item{cov_name}{A string indicating the name of the coverage to be retrieved.}
}
\value{
A spatial data frame containing the filtered data from the geoserver.
}
\description{
Downloads and filters spatial data from the geoserver for a specified tier and coverage name.
}
\details{
Requires BY_TIER, DATA_PATH, and geo_info to be available in the environment.
}
\examples{
get_geoserver_info()
cov_data <- get_geoserver_data(Tier = 4, cov_name = "degrees_heating_weeks_tier")
}
\author{
Murray Logan
}
