% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geoserver_info.R
\name{get_geoserver_info}
\alias{get_geoserver_info}
\title{get geoserver info}
\usage{
get_geoserver_info()
}
\description{
Retrieves metadata from the geoserver
}
\details{
Get data from geoserver
}
\author{
Murray Logan
}
