% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lag_cov.R
\name{get_lag_cov}
\alias{get_lag_cov}
\title{Get lagged disturbance value}
\usage{
get_lag_cov(yr, cov, tier5, .x)
}
\arguments{
\item{yr}{Adjusted year of the ecological survey}

\item{cov}{Extracted disturbance layer corresponding to the monitored years}

\item{tier5}{Spatial unit identifier}

\item{.x}{Column name from the disturbance data (e.g. severity or max variable)}
}
\value{
Numeric value of the lagged disturbance, or 0 if missing
}
\description{
Filter disturbance values to extract lag from a previous year
}
\examples{
get_lag_cov(yr = 2020, cov = cov_dhw, tier5 = 101, .x = "severity_dhw")
}
\author{
Julie Vercelloni
}
