% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legacy_data.R
\name{get_legacy_data}
\alias{get_legacy_data}
\title{Get legacy data}
\usage{
get_legacy_data()
}
\description{
Get legacy data
}
\details{
Unzip, retrieve and read legacy benthic data
}
\author{
Murray
}
