% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stages.R
\name{get_stages}
\alias{get_stages}
\title{Function}
\usage{
get_stages()
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
