% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tiers.R
\name{get_tiers}
\alias{get_tiers}
\title{Get tiers}
\usage{
get_tiers()
}
\description{
Get tiers
}
\details{
This function retrieves the tiers from the geoserver and saves them as spatial data
}
\author{
Murray
}
