% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_log.R
\name{initialise_log}
\alias{initialise_log}
\title{Initialise log}
\usage{
initialise_log()
}
\value{
creates log file\cr
and the globabal variable LOG_FILE, which contains file location
}
\description{
The following function initialises a log file. This log file is placed in the root of the project as it needs to be in a location that is guarenteed to exist from a freshly cloned instance of this codebase.
}
\examples{
initialise_log
}
