% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_prep.R
\name{inla_prep}
\alias{inla_prep}
\title{Function}
\usage{
inla_prep(data.grp.tier.ready, HexPred_reefid2)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples
}
