% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_covariates_to_tier_lookup.R
\name{join_covariates_to_tier_lookup}
\alias{join_covariates_to_tier_lookup}
\title{Join Covariates to Tier Lookup}
\usage{
join_covariates_to_tier_lookup(tier.sf)
}
\arguments{
\item{tier.sf}{A spatial (`sf`) layer containing Tier5 information.}
}
\value{
A spatial (`sf`) layer with joined covariates.
}
\description{
Merges spatial tier data with the tier lookup table to attach covariates.
}
\examples{
\dontrun{
tier.sf <- readRDS("data/primary/tier5.sf.RData")
tier.sf.joined <- join_covariates_to_tier_lookup(tier.sf)
}
}
\author{
Julie Vercelloni
}
