% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_for_model.R
\name{load_data_for_model}
\alias{load_data_for_model}
\title{Load Data for Model Execution}
\usage{
load_data_for_model()
}
\arguments{
\item{DATA_PATH}{Character. Path to the folder containing `processed/` and `primary/` data directories.}

\item{RDATA_FILE}{Character. Name of the `.RData` file to load from the `processed/` directory.}
}
\value{
No return value. Loads data into the environment
}
\description{
Loads benthic survey data and optionally identifies if covariates are available.
This function checks for required `.RData` files in the processed data directory,
sets up the list of target groups, and flags if covariate data is available.
}
\examples{
load_data_for_model()
}
