% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEST.R, R/loadRData.R
\name{loadRData}
\alias{loadRData}
\title{Function}
\usage{
loadRData(fileName)

loadRData(fileName)
}
\arguments{
\item{fileName}{.RData file name}

\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description

Reads RData, and assigns a name
}
\examples{
examples
loadRData(".RData")
}
