% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRData.R
\name{loadRData}
\alias{loadRData}
\title{loadRData}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{.RData file name}
}
\description{
Reads RData, and assigns a name
}
\examples{
loadRData(".RData")
}
