% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tiers_lookup.R
\name{make_tiers_lookup}
\alias{make_tiers_lookup}
\title{Function}
\usage{
make_tiers_lookup()
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
