% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitmodel.R
\name{model_fitModel}
\alias{model_fitModel}
\title{Fit model}
\usage{
model_fitModel(raw_cell_means = TRUE)
}
\arguments{
\item{raw_cell_means}{boolean. If TRUE, runs raw cell means. If FALSE, runs simple INLA cell means. Default is TRUE.}
}
\description{
Fits model to data
}
\examples{
model_fitModel()
}
