% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_ModelTier_type6.R
\name{model_fitModelTier_type6}
\alias{model_fitModelTier_type6}
\title{Fit model at tier level}
\usage{
model_fitModelTier_type6(data.grp.not.enough, tier.sf)
}
\arguments{
\item{data.grp.not.enough}{data with low replication at the tier level}

\item{tier.sf}{covariates shapefile}
}
\description{
Fits INLA model to data at tier level when the number of observations is low
}
\examples{
model_fitModelTier_type6()
}
\author{
Julie Vercelloni
}
