% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_loadData_unix.R
\name{model_loadData_unix}
\alias{model_loadData_unix}
\title{Load data}
\usage{
model_loadData_unix()
}
\description{
This script will copy all the data sources from the bucket provided and place in <DATA_PATH>/primary.\cr
When run on the dev server (HPC) DATA_FROM will be LOCAL\cr
When run on the prod server (AWS) DATA_FROM will be S3\cr
There should be three data collections in <AWS_PATH>/raw:\cr
- benthic data (reefs_data.csv|zip)\cr
- spatial tiers data (tiers[2-5].shp)\cr
- covariates data (?)
}
\examples{
model_loadData_unix()
}
