% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_processData.R
\name{model_processData}
\alias{model_processData}
\title{Process data}
\usage{
model_processData()
}
\description{
## This script predominantly loads the benthic data (including legacy
}
\examples{
model_processData()
}
