% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCLA.R
\name{parseCLA}
\alias{parseCLA}
\title{parseCLA}
\usage{
parseCLA(args)
}
\arguments{
\item{args}{command line arguments. The call must be of the form:
30_model.R --bucket="<PATH>"
--domain=<DOMAIN>
  [--by_tier=<NUMBER> --debug=<true|false> --runStage=<NUM> --refresh_data=<true|false>]


<PATH>:      a valid path to a folder containing the input data
<DOMAIN>:    either tier (spatio-temporal model for the data provide)
or site (hierarchical model for a single site.
         <NUMBER>:    for tier analyses, an optional tier number to indicate which
         tier level to use in splitting up analyses [n>true|false>:     whether to operate in debug mode.
          If true, progress is provided via a CLI
<NUM>:     which stages of the analysis to run (-1 or missing is all stages}
}
\value{
returned arguments description
}
\description{
The following function parses the command line arguments
1. Ensure that a bucket has been supplied as --bucket=<bucket>
2. If the bucket is local (/data/.* or /mnt/reefcloud/.*)
   - set DATA_FROM <<- "LOCAL"
   If the bucket is S3 (s3://dev-aims-gov-au-reefcloud-stats/reefcloud/.*)
   - set DATA_FROM <<- "S3"
3. Ensure that a valid domain has been supplied and if so
}
\examples{
examples
}
