% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCLA.R
\name{parseCLA}
\alias{parseCLA}
\title{parseCLA}
\usage{
parseCLA(args)
}
\arguments{
\item{args}{command line arguments. The call must be of the form:\cr
"--bucket="<PATH> --domain=<DOMAIN> --by_tier=<NUMBER> --debug=<true|false> --runStage=<NUM> --refresh_data=<true|false>"\cr
<PATH>:      a valid path to a folder containing the input data\cr
<DOMAIN>:    either tier (spatio-temporal model for the data provide) or site (hierarchical model for a single site.\cr
<NUMBER>:    for tier analyses, an optional tier number to indicate which tier level to use in splitting up analyses \cr
<true|false>:     whether to operate in debug mode. If true, progress is provided via a CLI\cr
<NUM>:     which stages of the analysis to run (-1 or missing is all stages)}
}
\value{
Creates the following global variables \cr
REFRESH_DATA\cr
DEBUG_MODE\cr
runstage\cr
runStage\cr
AWS_PATH\cr
DATA_FROM\cr
DOMAIN_CATEGORY \cr
AWS_OUTPUT_PATH \cr
DOMAIN_NAME \cr
GENERATE_REPORT \cr
MODEL_TYPE \cr
BY_TIER
}
\description{
The following function parses the command line arguments \cr
1. Ensure that a bucket has been supplied as --bucket=<bucket>\cr
2. If the bucket is local (/data/.* or /mnt/reefcloud/.*)\cr
   - set DATA_FROM <<- "LOCAL"\cr
   If the bucket is S3 (s3://dev-aims-gov-au-reefcloud-stats/reefcloud/.*)\cr
   - set DATA_FROM <<- "S3"\cr
3. Ensure that a valid domain has been supplied and if so\cr
}
\examples{
args = c("--bucket=/home/data/AUS", "--domain=tier", "--by_tier=4", "--debug=true", "--runStage=1", "--refresh_data=false")
parseCLA(args)
}
