% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_group_data_for_modelling.R
\name{prep_group_data_for_modelling}
\alias{prep_group_data_for_modelling}
\title{Group data for model}
\usage{
prep_group_data_for_modelling(data, GROUP)
}
\arguments{
\item{data}{data set}

\item{GROUP}{string, name of group}
}
\value{
grouped data, arranged by Tier4, Tier5, Site, Transect, and descending fYEAR
}
\description{
Prepare fGROUP for modelling
}
\examples{
prep_group_data_for_modelling(data, GROUP)
}
