% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_group_data_for_modelling.R
\name{prep_group_data_for_modelling}
\alias{prep_group_data_for_modelling}
\title{Function}
\usage{
prep_group_data_for_modelling(data, GROUP)
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
