% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data for Analyses}
\usage{
prepare_data(data)
}
\arguments{
\item{data}{A data frame containing the benthic data with required fields (e.g., REPORT_YEAR, GROUP_DESC, SITE_DEPTH, etc.).}
}
\value{
A cleaned and formatted `data.frame` ready for analysis and reporting.
}
\description{
Processes and summarizes benthic monitoring data, including grouping, aggregation, and formatting required for downstream analysis. Saves both `.RData` and `.csv` outputs.
}
\examples{
\dontrun{
load("raw_benthic_data.RData")
processed_data <- prepare_data(benthic_data)
}
}
\author{
Murray Logan
}
