% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud__add_stage.R
\name{ReefCloud__add_stage}
\alias{ReefCloud__add_stage}
\title{ReefCloud__add_stage}
\usage{
ReefCloud__add_stage(stage, title)
}
\arguments{
\item{stage}{stage id (integer)}

\item{title}{name of stage (character)}
}
\value{
The global variable STATUS is modified
}
\description{
Adds stage to STATUS global variable
}
\examples{
examples
}
