% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReefCloud_startMatter.R
\name{ReefCloud_startMatter}
\alias{ReefCloud_startMatter}
\title{Function}
\usage{
ReefCloud_startMatter(args = commandArgs())
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
