% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_benthic_data.R
\name{retrieve_benthic_data}
\alias{retrieve_benthic_data}
\title{Retrieve benthic data}
\usage{
retrieve_benthic_data()
}
\value{
NULL (used for side effects: copying or downloading files)
}
\description{
Retrieve the benthic data from the source specified by `AWS_PATH` and `FILENAME`, and save it to the `DATA_PATH` directory
}
\examples{
retrieve_benthic_data()
}
\author{
Murray Logan
}
