% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_benthic_data.R
\name{save_benthic_data}
\alias{save_benthic_data}
\title{Save benthic data}
\usage{
save_benthic_data(data)
}
\arguments{
\item{data}{A data frame or list to be saved}
}
\value{
NULL (used for side effects: writing files and updating state)
}
\description{
Save benthic data to an RData file in the primary data directory. Optionally updates the analysis stage.
}
\examples{
save_benthic_data(benthic_df)
}
\author{
Murray
}
