% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_status.R
\name{save_status}
\alias{save_status}
\title{Function}
\usage{
save_status()
}
\arguments{
\item{parameters}{description}
}
\value{
returned arguments description
}
\description{
Description
}
\examples{
examples 
}
