% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_benthic_data}
\alias{validate_benthic_data}
\title{Validate benthic data}
\usage{
validate_benthic_data(data, rules)
}
\arguments{
\item{data}{Benthic data}

\item{rules}{Rules to validate the data}
}
\value{
validation result or the violating rows of the benthic data
}
\description{
Validate benthic data
}
\details{
Validate benthic data
}
\author{
Murray
}
